+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Filter down ticks on horizontal `<BarChart />` based on chart size.
+
 ## [0.28.2] - 2021-12-13
 
 - Replaced `theme.line.dottedStrokeColor` by `theme.seriesColors.comparison`